Use Int value of chainId in signature

+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/models/Signature.kt
@@ -2,7 +2,7 @@ package io.horizontalsystems.ethereumkit.models
 
 import io.horizontalsystems.ethereumkit.core.toHexString
 
-class Signature(val v: Byte,
+class Signature(val v: Int,
                 val r: ByteArray,
                 val s: ByteArray) {
     override fun toString(): String {
